refactor: 테스트 통합 및 Qualifier App 모듈에서 정의
