+++ app/src/main/java/woowacourse/shopping/di/InMemory.kt
@@ -1,4 +1,6 @@
-package com.di.berdi.annotation
+package woowacourse.shopping.di
+
+import com.di.berdi.annotation.Qualifier
 
 @Qualifier
 @Retention(AnnotationRetention.RUNTIME)
