+++ app/src/main/java/woowacourse/shopping/di/NormalModule.kt
@@ -3,8 +3,6 @@ package woowacourse.shopping.di
 import android.content.Context
 import androidx.room.Room
 import com.di.berdi.Module
-import com.di.berdi.annotation.InMemory
-import com.di.berdi.annotation.OnDisk
 import woowacourse.shopping.data.CartInMemoryRepository
 import woowacourse.shopping.data.CartOnDiskRepository
 import woowacourse.shopping.data.CartProductDao
@@ -24,12 +22,15 @@ object NormalModule : Module {
 
     fun provideProductRepository(): ProductRepository = ProductDefaultRepository()
 
-    fun provideCartProductDao(context: Context): CartProductDao {
-        val db = Room.databaseBuilder(
+    fun provideCartProductDao(database: ShoppingDatabase): CartProductDao {
+        return database.cartProductDao()
+    }
+
+    fun provideShoppingDB(context: Context): ShoppingDatabase {
+        return Room.databaseBuilder(
             context.applicationContext,
             ShoppingDatabase::class.java,
             "shopping_db",
         ).build()
-        return db.cartProductDao()
     }
 }
