+++ app/src/main/java/woowacourse/shopping/ui/MainViewModel.kt
@@ -4,8 +4,8 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.di.berdi.annotation.OnDisk
 import kotlinx.coroutines.launch
+import woowacourse.shopping.di.OnDisk
 import woowacourse.shopping.model.Product
 import woowacourse.shopping.repository.CartRepository
 import woowacourse.shopping.repository.ProductRepository
