+++ app/src/test/java/woowacourse/shopping/di/diactivity/DIActivityTest.kt
@@ -5,9 +5,7 @@ import com.di.berdi.Container
 import com.di.berdi.DIActivity
 import com.di.berdi.DIApplication
 import com.di.berdi.Injector
-import com.di.berdi.annotation.InMemory
 import com.di.berdi.annotation.Inject
-import com.di.berdi.annotation.OnDisk
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -16,64 +14,42 @@ import org.robolectric.RobolectricTestRunner
 import org.robolectric.annotation.Config
 import woowacourse.shopping.di.FakeObj
 import woowacourse.shopping.di.FakeRepository
+import woowacourse.shopping.di.InMemory
 import woowacourse.shopping.di.InMemoryFakeRepository
+import woowacourse.shopping.di.OnDisk
 import woowacourse.shopping.di.OnDiskFakeRepository
 
-@RunWith(RobolectricTestRunner::class)
-class DIActivityTest {
-    inner class FakeAnnotationDIApplication : DIApplication() {
-        override fun inject() {
-            val container = Container()
-            injector =
-                Injector(container, applicationContext).apply { inject(FakeObjModule) }
-        }
-    }
-
-    @Test
-    @Config(application = FakeAnnotationDIApplication::class)
-    fun `어노테이션을 사용한 프로퍼티만 주입된다`() {
-        class FakeAnnotationDIViewModel : ViewModel() {
-            @Inject
-            lateinit var fakeObj: FakeObj
-        }
-
-        class FakeAnnotationDIActivity : DIActivity() {
-            lateinit var viewModel: FakeAnnotationDIViewModel
-        }
-
-        val activity = Robolectric
-            .buildActivity(FakeAnnotationDIActivity::class.java)
-            .create()
-            .get()
-
-        val viewModel = activity?.viewModel!!
-
-        assertThat(viewModel.fakeObj).isSameAs(FakeObj)
-    }
-
-    inner class FakeConstructorDIApplication : DIApplication() {
-        override fun inject() {
-            val container = Container()
-            injector = Injector(container, applicationContext).apply {
-                inject(FakeRepositoryModule)
+class FakeDIApplication : DIApplication() {
+    override fun inject() {
+        injector =
+            Injector(Container(), applicationContext).apply {
+                injectBy(
+                    listOf(FakeRepositoryModule, FakeObjModule),
+                )
             }
-        }
     }
+}
+
+class FakeDIViewModel(
+    @InMemory val inMemoryFakeRepository: FakeRepository,
+    @OnDisk val onDiskFakeRepository: FakeRepository,
+) : ViewModel() {
+    @Inject
+    lateinit var fakeObj: FakeObj
+}
+
+@RunWith(RobolectricTestRunner::class)
+@Config(application = FakeDIApplication::class)
+class DIActivityTest {
 
     @Test
-    @Config(application = FakeConstructorDIApplication::class)
     fun `맞는 타입의 객체 인스턴스가 존재하면 ViewModel 의존성 주입이 성공한다`() {
-        class FakeConstructorDIViewModel(
-            @InMemory val inMemoryFakeRepository: FakeRepository,
-            @OnDisk val onDiskFakeRepository: FakeRepository,
-        ) : ViewModel()
-
-        class FakeConstructorDIActivity : DIActivity() {
-            lateinit var viewModel: FakeConstructorDIViewModel
+        class FakeDIActivity : DIActivity() {
+            lateinit var viewModel: FakeDIViewModel
         }
 
         val activity = Robolectric
-            .buildActivity(FakeConstructorDIActivity::class.java)
+            .buildActivity(FakeDIActivity::class.java)
             .create()
             .get()
 
@@ -84,5 +60,8 @@ class DIActivityTest {
         // and
         assertThat(viewModel?.inMemoryFakeRepository).isInstanceOf(InMemoryFakeRepository::class.java)
         assertThat(viewModel?.onDiskFakeRepository).isInstanceOf(OnDiskFakeRepository::class.java)
+
+        // and: field inject
+        assertThat(viewModel?.fakeObj).isSameAs(FakeObj)
     }
 }
