+++ app/src/test/java/woowacourse/shopping/di/diactivity/FakeRepositoryModule.kt
@@ -1,13 +1,13 @@
 package woowacourse.shopping.di.diactivity
 
 import com.di.berdi.Module
-import com.di.berdi.annotation.InMemory
-import com.di.berdi.annotation.OnDisk
 import woowacourse.shopping.di.DefaultFirstDataSource
 import woowacourse.shopping.di.DefaultSecondDataSource
 import woowacourse.shopping.di.FakeRepository
 import woowacourse.shopping.di.FirstDataSource
+import woowacourse.shopping.di.InMemory
 import woowacourse.shopping.di.InMemoryFakeRepository
+import woowacourse.shopping.di.OnDisk
 import woowacourse.shopping.di.OnDiskFakeRepository
 import woowacourse.shopping.di.SecondDataSource
